Cleanup and license the dongle drivers code

+++ mbed-dongle-firmware/drivers/glove_sensors.h
@@ -13,10 +13,32 @@
  *  This is the definition of the struct used on either end
  *  of the bluetooth connection, and it contains synchronization things.
  *
+ * Copyright (c) 2016 by Nick Bertoldi, Ben Heckathorn, Ryan O'Keefe,
+ *                       Adrian Padin, Timothy Schumacher
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the "Software"),
+ * to deal in the Software without restriction, including without limitation
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+ * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+ * IN THE SOFTWARE.
  */
 
 #ifndef GLOVE_SENSORS_H_
 #define GLOVE_SENSORS_H_
+
+#include <inttypes.h>
 
 #include "flex_sensor.h"
 #include "touch_sensor.h"
@@ -37,8 +59,8 @@ typedef struct {
  * compressed version of all the
  * sensor data that fits in a BLE advertisement
  */
-const size_t glove_sensors_compressed_size = 22;
-const size_t glove_sensors_compressed_size_no_crc = 20;
+const uint8_t glove_sensors_compressed_size = 22;
+const uint8_t glove_sensors_compressed_size_no_crc = 20;
 typedef struct {
     // flex sensors (12*4 = 48b = 6B)
     uint16_t f[3];
